1st deploy-version

+++ routes/index.js
@@ -5,7 +5,7 @@ const User = require('../models/User');
 const cloudinary = require('../config/cloudinary');
 const nodemailer = require('../config/nodemailer');
 
-module.exports = function(app, passport, newsapi) {
+module.exports = function(app, passport, newsapi,mongoose) {
   app.get("/", (req, res, next) => {
     newsapi.v2.topHeadlines({ country: "us" })
       .then((response)=>{
@@ -100,17 +100,15 @@ module.exports = function(app, passport, newsapi) {
   });
 
   app.post('/password/recover/:id', async (req, res, next) => {
-    const { oldPassword, newPassword, newPassword2 } = req.body
+    const { newPassword, newPassword2 } = req.body
     const id = req.params.id;
-
-    console.log(req.body, id)
 
     try {
       const userDB = await User.findById(id)
 
       if(userDB.recover) {
         console.log('There is a recover code')
-        if(userDB.validatePassword(oldPassword) && newPassword === newPassword2) {
+        if(newPassword === newPassword2) {
           userDB.password = userDB.generateHash(newPassword);
           delete userDB.recover;
 
@@ -230,6 +228,11 @@ module.exports = function(app, passport, newsapi) {
           } else {
             art.liked = false;
           }
+        art.comments.forEach((eComment)=>{
+          if(eComment.user.equals(req.user._id)){
+            eComment.isMine = true;
+          }
+        })
         });
         if (req.user.following.includes(req.params.otherUserId)) {
           isFollowing = true;
@@ -499,6 +502,7 @@ module.exports = function(app, passport, newsapi) {
         articleUrl,
         articleDate
       });
+      await User.findByIdAndUpdate(req.user._id, {$push: {news: newArticle._id}})
     } catch (err) {
       console.log(err);
     }
@@ -511,9 +515,18 @@ module.exports = function(app, passport, newsapi) {
   });
 
   app.post("/news/delete", (req, res, next) => {
-    News.findOneAndRemove({ title: req.body.title })
+    console.log(req.body.title, req.user._id);
+    News.findOneAndRemove({$and: [{title: req.body.title },{owner: req.user._id}]})
       .then(response => {
-        res.json(response);
+        // console.log(response);
+        User.findByIdAndUpdate(req.user._id, {$pull: {news: response._id}})
+        .then((response)=>{
+          console.log(response);
+          res.json(response);
+        })
+        .catch((err)=>{
+          res.json(err);
+        })
       })
       .catch(err => {
         res.json(err);
@@ -533,10 +546,11 @@ module.exports = function(app, passport, newsapi) {
 
       getNew.comments.push(newComment._id);
       await getNew.save();
-
+      console.log(newComment.createdAt)
       res.json({
         ok: true,
         data: {
+          commentId: newComment._id,
           user: req.user,
           message: "Comment submitted"
         }
@@ -548,6 +562,71 @@ module.exports = function(app, passport, newsapi) {
       });
     }
   });
+  app.get("/checkfeed",async(req,res,next)=>{
+    try{
+      let allNews = [];
+      const me = await User.findById(req.user._id);
+      let people = me.following;
+      let allUsers = await User.find()
+      let news = await News.find().populate({
+        path: "comments",
+        populate: { path: "user" },
+      })
+     news.forEach((eachNews)=>{
+       people.forEach((ePerson)=>{
+         if(eachNews.owner.equals(ePerson)){
+           allNews.push(eachNews);
+         }
+       })
+     })
+     allNews.forEach((eachNews)=>{
+       allUsers.forEach((ePerson)=>{
+         if(eachNews.owner.equals(ePerson._id)){
+           eachNews.userImage = ePerson.img;
+           eachNews.userName = ePerson.name;
+           eachNews.userUserName = ePerson.username;
+         }
+       })
+     })
+      // let sorted = allNews.sort((a,b)=>{
+      //   if(a.createdAt > b.createdAt){
+      //     return -1;
+      //   }
+      //   if(a.createdAt < b.createdAt){
+      //     return 1;
+      //   }
+      //   return 0;
+      // })
+      // console.log(allNews);
+      // let allNews = await News.find().sort({createdAt: -1}).populate("owner").populate({
+      //   path: "comments",
+      //   populate: { path: "user" }
+      // }).limit(50);
+      
+      // console.log(allNews);
+      res.render('yourFeed',{allNews: allNews, user: req.user, comments: allNews.comments})
+    }catch(err){
+      console.log(err);
+    }
+  })
+  app.post("/comment/delete", async (req,res,next)=>{
+    const id = req.body.id;
+    try{
+      const comment = await Comment.findByIdAndDelete(id);
+      let articleId = comment.article;
+      console.log(articleId)
+      const newArticle = await News.findByIdAndUpdate(articleId, {$pull: {comments: id}})
+      console.log(newArticle);
+      res.json({
+        ok: true
+      })
+    }catch (err) {
+      res.json({
+        ok: false,
+        data: "Something went wrong"
+      });
+    }
+  })
 
   app.post("/likes/add", async (req, res) => {
     const { title } = req.body;
